iso: support for per-device title scanning function, improve title ID cache by storing only titles present on the device
